Adds tint, adds image content, removes unnecessary image content, resizes images, cleans up code, adds active class to circle button.
Added different background colour to circle-button class to show that
the page is currently being accessed. Removed page card for page the
user is already on.

Added tinting of bg image on wrapper for pages one, two and three
Generally reduced sizes of images to help with load speed.

Deleted unnecessary lines of code.